Refactor generate_targets into separate module with state consolidated into new TargetGenerator struct
